Fix localStorage/permalink bug
Fix a bug where the map directive converts x, y and zoom from version 2 to version 3 when the values are read from the local storage. There should be no conversion when the values are read from the local storage. The conversions occur only when reading from the location URL, when there's no version parameter or when the version parameter is set to 2.

Fixes #372.