Use template cache also in debug mode
This commit changes the `Makefile` and the `index.html` template in a way that the Angular template cache (`templatecache.js`) is used in non-debug mode as well as in debug mode. Two benefits: (1) fix a bug where the theme switcher collapse/expand doesn't work correctly in debug mode (#826), and (2) make the debug and non-debug modes closer, thereby avoiding surprises.