Prevent property renaming in print directive
The `layout` and `scale` properties of `app.PrintController` are defined using the bracket notation. For example, this is how the `layout` property is defined in the constructor:

```js
/**
 * @type {string}
 */
this['layout'] = this['layouts'][0];
```

So when accessing these properties the bracket notation must be used as well. Do not ever mix the bracket and dot notations for a given property.