Disable 3D by default
Having 3D disabled by default makes it possible to merge 3D code in
master without impact to the production.

The 3D code will be dormant until the project is complete and the features
go live.

Being able to merge regularly to master avoid the two branches to diverge
and share common work like updating the dependencies and modernizing
the code.