Exaggerate terrain and load all terrain
Exaggerating terrain requires finer terrain. The previous terrain optimization
(loading only a subset of the terrain) was producing artefacts with a terrain
exaggeration of 2; it is now disabled by default.

2 parameters can now be passed in the URL:

- terrain_exaggeration=float
The terrain will be exaggerated of that value. When the parameter is
missing, defaults to 2.0.

- terrain_levels=<coma separated list of zoom levels>
The terrain levels dowloaded by Cesium will be restricted to the
provided ones. When parameter is missing, defaults to all.
Example: "terrain_levels=8,11,14,16,17" (old default).