Handle different clients
Different clients (window/tab) have a different FetchEvent.clientId property.
We maintain a set of offline clients in the service worker. When a client wants to
switch to offline mode a query is sent to the worker.

iOS/Safari do not support the FetchEvent.clientId property so there is no way to track which client is offline.
As a consequence the last loaded/offlined tab or window wins. Other clients are in an invalid state.