Fix ngeo dep
We where previously using a custom build of ngeo for lux. But it's not
necessary anymore (like a custom build of ol-cesium but that's cover
with another commit). I also removed raven-js which is the old sentry
agent.

Also fix DB connection and `docker-compose.override.sample.yaml`
indentation of the db service declaration